updated settings.py for hadoop job limit.
turned the hadoop job count down to 50 so that it doesn't take so long for a user submission to be get addressed by the system.  With it set to 150, it took about 6 hours for an item submitted to first receive service.  This should get that latency down so long as it doesn't exclude the number of users we have at the front of the line.  Will have to keep an eye on it and possibly adjust later.